(feat): Cells now can move when clicking arrows
A biggest commit yet. Now cells move when you
click on one of the arrows, and a state of the
Game component changes after a whole move, not
just one cell moving. Also had to change "shift" classes in .sass. Also added a "cells"
property inside Game's state, so that the board
rerenders after a cell spawn.